+++ src/Facility.Core/Http/JsonHttpContentSerializerSettings.cs
@@ -1,3 +1,6 @@
+using System;
+using System.IO;
+
 namespace Facility.Core.Http
 {
 	/// <summary>
@@ -9,5 +12,11 @@ namespace Facility.Core.Http
 		/// True to force async I/O, even if a large memory buffer is required.
 		/// </summary>
 		public bool ForceAsyncIO { get; set; }
+
+		/// <summary>
+		/// Called to create a memory stream. Defaults to creating a new <see cref="MemoryStream" />.
+		/// </summary>
+		/// <remarks>Consider using <c>Microsoft.IO.RecyclableMemoryStream</c>.</remarks>
+		public Func<Stream>? MemoryStreamCreator { get; set; }
 	}
 }